a few context tests
